Refactors minimizer structure and imports
