+++ tests/unit/analysis/minimizers/test_minimizer_base.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 
 import pytest
 
-from easydiffraction.analysis.minimizers.minimizer_base import FitResults
+from easydiffraction.analysis.fitting.results import FitResults
 from easydiffraction.analysis.minimizers.minimizer_base import MinimizerBase
